requirements.txt
aioftp==0.27.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
beautifulsoup4==4.14.3
cachetools==5.5.2
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
decorator==5.2.1
dnspython==2.8.0
exceptiongroup==1.3.1
fastapi==0.135.3
greenlet==3.4.0
h11==0.16.0
holidays==0.94
idna==3.11
il-supermarket-parser==0.2.2
il-supermarket-scraper==1.0.0
iniconfig==2.3.0
lxml==5.2.1
mock==4.0.3
numpy==2.2.6
packaging==26.0
pandas==2.2.2
playwright==1.58.0
pluggy==1.6.0
py==1.11.0
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.1
python-dotenv==1.0.1
Pygments==2.20.0
pymongo==4.6.3
pytest==9.0.3
pytest-base-url==2.1.0
pytest-playwright==0.7.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2026.1.post1
requests==2.33.1
retry==0.9.2
six==1.17.0
soupsieve==2.8.3
starlette==1.0.0
text-unidecode==1.3
tomli==2.4.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
uvicorn==0.44.0
